+++ Sources/MapboxMaps/SwiftUI/Deps.swift
@@ -13,7 +13,9 @@ struct MapDependencies {
     var eventsSubscriptions = [AnyEventSubscription]()
     var cameraChangeHandlers = [(CameraChanged) -> Void]()
     var ornamentOptions = OrnamentOptions()
+    var frameRate = Map.FrameRate()
     var debugOptions = MapViewDebugOptions()
+    var isOpaque = true
     var presentsWithTransaction = false
     var additionalSafeArea = SwiftUI.EdgeInsets()
     var viewportOptions = ViewportOptions(transitionsToIdleUponUserInteraction: true, usesSafeAreaInsetsAsPadding: true)
@@ -45,5 +47,13 @@ extension Map {
     struct PerformanceStatisticsParameters {
         var options: PerformanceStatisticsOptions
         var callback: (PerformanceStatistics) -> Void
+    }
+}
+
+@available(iOS 13.0, *)
+extension Map {
+    struct FrameRate: Equatable {
+        var range: ClosedRange<Float>?
+        var preffered: Float?
     }
 }